j2discordbot/requirements.txt
discord.py
sleekxmpp
